Introduce ContentParser for content rendering
This introduces a new `ContentParser` class that allows the ability to
register various "transformers".

Here I've registered a Markdown transformer, as well as a transformer
that can scan the content for GitHub user names and convert them to
links.

The idea is that this is totally extensible and stackable, and we can
easily introduce any extra transformations we want down the road.